rename user route file
